fix: use np.float32 because of np.float16 issues
ref: https://github.com/scipy/scipy/issues/7408

+++ src/problems/tsp/tsp_ga.py
@@ -52,7 +52,7 @@ class TSPGAProblem(Problem):
             shape = (self.instance.n * self.config.sparse_factor,)
 
         for i in range(values.shape[0]):
-            random_heatmap = np.ones(shape, dtype=np.float16) if i == 0 else np.random.rand(*shape).astype(np.float16)
+            random_heatmap = np.ones(shape, dtype=np.float32) if i == 0 else np.random.rand(*shape).astype(np.float32)
             values[i] = self.instance.get_tour_from_adjacency_np_heatmap(random_heatmap)
 
     def _fill_difusco_sampling(self, values: torch.Tensor) -> None:
